Task: 0. Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,13 +11,13 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i;
+	size_t i;
 
 	if (!array)
 		return (-1);
 	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%i] = [%i]\n", i, array[i]);
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
